bulld updates, built and tests passes for local interfacing with lighthouse

+++ src/lib.rs
@@ -5,9 +5,9 @@ pub mod simulator;
 pub mod validation;
 
 
-pub type DvfCommitteeIndex = u32;
+pub type DvfCommitteeIndex = u64;
 
 /// The id of the operator in a threshold signature instance 
-pub type DvfOperatorTsid = u32;
+pub type DvfOperatorTsid = u64;
